Code updated to display runs as they happen.
Had to update the way song data is stored to be able to properly process a run.

+++ core/src/com/game/helper/SongDataFileLoader.java
@@ -35,7 +35,7 @@ public class SongDataFileLoader
          */
         public int[] types;
         
-        public int[][] runs;
+        public float[][] runs;
     }
     
     /**
